Update card_automation.py
For cards that allow the player to choose another investigator, i added 2 functions (InvestigatorColor and InvestigatorName) to allow the customization of the choice_list menus.

AttachTo function is used for some cards like Sefina, Ancestral Knowledge, Stick to the Plan, Backpack and such

Regrouped the list in Guardian, Seeker, Mystic, Rogue and Survivor for clarity.